requirements.txt
aiogram==2.4
Pillow